Remove preventAnnouncement and move block description up to the grid cell level
