+++ packages/block-editor/src/components/list-view/block-select-button.js
@@ -6,17 +6,14 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
-import { Button, VisuallyHidden } from '@wordpress/components';
-import { useDebounce, useInstanceId, usePrevious } from '@wordpress/compose';
-import { forwardRef, useEffect, useState } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
+import { Button } from '@wordpress/components';
+import { forwardRef } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import BlockIcon from '../block-icon';
 import useBlockDisplayInformation from '../use-block-display-information';
-import { getBlockPositionDescription } from './utils';
 import BlockTitle from '../block-title';
 import ListViewExpander from './expander';
 import { SPACE, ENTER } from '@wordpress/keycodes';
@@ -25,35 +22,17 @@ function ListViewBlockSelectButton(
 	{
 		className,
 		block: { clientId },
-		isSelected,
 		onClick,
 		onToggleExpanded,
-		position,
-		siblingBlockCount,
-		level,
 		tabIndex,
 		onFocus,
 		onDragStart,
 		onDragEnd,
 		draggable,
-		isExpanded,
-		preventAnnouncement,
 	},
 	ref
 ) {
 	const blockInformation = useBlockDisplayInformation( clientId );
-	const instanceId = useInstanceId( ListViewBlockSelectButton );
-	const descriptionId = `list-view-block-select-button__${ instanceId }`;
-	const blockPositionDescription = getBlockPositionDescription(
-		position,
-		siblingBlockCount,
-		level
-	);
-	const [ ariaHidden, setAriaHidden ] = useState( undefined );
-
-	// This debounced version is used so that while moving out of focus,
-	// the block isn't updated and then re-announced.
-	const delaySetAriaHidden = useDebounce( setAriaHidden, 200 );
 
 	// The `href` attribute triggers the browser's native HTML drag operations.
 	// When the link is dragged, the element's outerHTML is set in DataTransfer object as text/html.
@@ -70,21 +49,6 @@ function ListViewBlockSelectButton(
 		}
 	}
 
-	const previousPreventAnnouncement = usePrevious( preventAnnouncement );
-
-	useEffect( () => {
-		// If we prevent screen readers from announcing the block,
-		// we should apply this immediately.
-		if ( preventAnnouncement ) {
-			setAriaHidden( true );
-		}
-		// Delay re-enabling so that if focus is being moved between
-		// buttons, we don't accidentally re-announce a focused button.
-		if ( ! preventAnnouncement && previousPreventAnnouncement ) {
-			delaySetAriaHidden( undefined );
-		}
-	}, [ preventAnnouncement ] );
-
 	return (
 		<>
 			<Button
@@ -94,7 +58,6 @@ function ListViewBlockSelectButton(
 				) }
 				onClick={ onClick }
 				onKeyDown={ onKeyDownHandler }
-				aria-describedby={ descriptionId }
 				ref={ ref }
 				tabIndex={ tabIndex }
 				onFocus={ onFocus }
@@ -102,8 +65,7 @@ function ListViewBlockSelectButton(
 				onDragEnd={ onDragEnd }
 				draggable={ draggable }
 				href={ `#block-${ clientId }` }
-				aria-expanded={ isExpanded }
-				aria-hidden={ ariaHidden }
+				aria-hidden={ true }
 			>
 				<ListViewExpander onClick={ onToggleExpanded } />
 				<BlockIcon icon={ blockInformation?.icon } showColors />
@@ -113,18 +75,7 @@ function ListViewBlockSelectButton(
 						{ blockInformation.anchor }
 					</span>
 				) }
-				{ isSelected && (
-					<VisuallyHidden>
-						{ __( '(selected block)' ) }
-					</VisuallyHidden>
-				) }
 			</Button>
-			<div
-				className="block-editor-list-view-block-select-button__description"
-				id={ descriptionId }
-			>
-				{ blockPositionDescription }
-			</div>
 		</>
 	);
 }
